+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # Changelog
 # v0.9.?(2023-02-14)
 * Breaking changes, mappings.yml refactored 
-* Added `parallel` property to config, which executes different sources in threads.
+* Added `threads` property to config, which executes different sources in threads.
 * WebUI: Added clipboard collector on left side 
 
 # v0.9.6(2023-01-14)